Added "validations"  for menu-option 1 & 2
